requirements.txt
flask==3.0.0
gunicorn==21.2.0
yfinance>=0.2.66
pandas>=2.2.0
numpy>=1.26.0
ta==0.11.0
vaderSentiment==3.3.2
requests==2.32.5
beautifulsoup4==4.14.3
feedparser==6.0.10
python-dotenv==1.0.0
scikit-learn>=1.4.0
joblib>=1.3.0
PyPDF2>=3.0.0
google-generativeai>=0.3.0
praw>=7.7.0
finvizfinance>=0.14.6
cloudscraper>=1.2.71
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

